+++ scripts/gulp/default.js
@@ -1,0 +1,11 @@
+'use strict';
+
+var gulp = require( 'gulp' );
+var runSequence = require('run-sequence');
+
+gulp.task('default', function(callback) {
+  runSequence(
+    ['build'],
+    ['test']
+  );
+});