+++ scripts/gulp/parseComponentName.js
@@ -1,0 +1,4 @@
+var argv = require('minimist')(process.argv.slice(2)),
+    component = argv.component || argv.c;
+
+module.exports = component;