+++ scripts/npm/prepublish/lib/getNpmVersion.js
@@ -1,0 +1,12 @@
+var RegClient = require('silent-npm-registry-client'),
+    client = new RegClient(),
+    promisify = require('promisify-node');
+
+promisify(client);
+
+function getVersion(component) {
+  var uri = 'https://registry.npmjs.org/' + component;
+  return client.get(uri, {timeout: 1000});
+}
+
+module.exports = getVersion;