+++ scripts/npm/prepublish/lib/gitStatus.js
@@ -1,0 +1,7 @@
+var exec = require('child-process-promise').exec;
+
+function status(path) {
+  return exec('git status -s ' + path);
+}
+
+module.exports = status;